BackEnd/voicetraining/voicetrain/requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
dialogflow==1.1.0
Django==3.1.3
djangorestframework==3.12.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-speech==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
jmespath==0.10.0
libcst==0.3.14
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
py-hanspell==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
